+++ backend/libraries/ledger_utils/src/lib.rs
@@ -56,7 +56,7 @@ pub fn default_ledger_account(principal: Principal) -> AccountIdentifier {
 }
 
 pub fn convert_to_subaccount(principal: &Principal) -> Subaccount {
-    let mut subaccount = [0; std::mem::size_of::<Subaccount>()];
+    let mut subaccount = [0; size_of::<Subaccount>()];
     let bytes = principal.as_slice();
     subaccount[0] = bytes.len().try_into().unwrap();
     subaccount[1..1 + bytes.len()].copy_from_slice(bytes);