+++ frontend/openchat-agent/src/services/user/candid/types.d.ts
@@ -2274,6 +2274,7 @@ export interface SwapTokensArgs {
   'swap_id' : bigint,
   'input_token' : TokenInfo,
   'exchange_args' : { 'Sonic' : ExchangeArgs } |
+    { 'KongSwap' : ExchangeArgs } |
     { 'ICPSwap' : ExchangeArgs },
   'output_token' : TokenInfo,
 }