+++ backend/canisters/local_group_index/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Removed
+
+- Remove `push_events_v2` which is no longer used ([#6502](https://github.com/open-chat-labs/open-chat/pull/6502))
+
 ## [[2.0.1364](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1364-local_group_index)] - 2024-10-02
 
 ### Changed